--- a/.github/workflows/docker.yml
+++ b/.github/workflows/docker.yml
@@ -21,10 +21,10 @@ jobs:
     runs-on: ubuntu-latest
     steps:
 
-    - name: Set up Go 1.21
+    - name: Set up Go 1.22
       uses: actions/setup-go@v5
       with:
-        go-version: "1.21"
+        go-version: "1.22"
         cache: false
       id: go
 
@@ -89,10 +89,10 @@ jobs:
     runs-on: ubuntu-latest
     steps:
 
-    - name: Set up Go 1.21
+    - name: Set up Go 1.22
       uses: actions/setup-go@v5
       with:
-        go-version: "1.21"
+        go-version: "1.22"
         cache: false
       id: go