@majst01 2024-06-12
This will not be merged because the Issue reported was because the benchmark is doing two loops inside each other which is  responsible for a exponential increase in ip allocation time measured.